pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ds_fastapi"
dynamic = ["version"]
description = "A Python package for FastAPI. Pacakge is part of the greater GraspDP / DPW project."
dependencies = [
    "boto3",
    "fastapi",
    "requests",
    "pyjwt",
    "cryptography"
]
requires-python = ">=3.8"
license = {file = "LICENSE"}
readme = "PyPI.md"
keywords = ["grasp-labs", "python", "fastApi", "middleware", "aws", "authentication"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
authors = [
    {name = "yuan1989", email = "yuan@grasplabs.no"},
    {name = "skerve", email = "arve.skjervheim@gmail.com"},
]
maintainers = [
    {name = "skerve", email = "arve.skjervheim@gmail.com"},
]
[project.urls]
Homepage = "https://github.com/grasp-labs/ds-fastapi"
Documentation = "https://grasp-labs.github.io/ds-fastapi/"
Repository = "https://github.com/grasp-labs/ds-fastapi"
Issues = "https://github.com/grasp-labs/ds-fastapie/issues"
Changelog = "https://github.com/grasp-labs/ds-fastapi/blob/main/CHANGELOG.md"

[tool.setuptools.dynamic]
version = {file = "VERSION.txt"}
